split handlers into their own files and keep route additions with handlers, updated uid key to microid-componentid, added boiler plate cmd handler file
